requirements.txt
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
choreographer==1.0.10
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
defusedxml==0.7.1
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.20.1
et_xmlfile==2.0.0
fonttools==4.59.2
fpdf2==2.8.4
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.30.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaleido==1.0.0
kiwisolver==1.4.9
logistro==1.1.0
lxml==6.0.1
mariadb==1.1.13
MarkupSafe==3.0.2
matplotlib==3.10.6
narwhals==2.4.0
numpy==2.3.3
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pefile==2023.2.7
pillow==11.3.0
plotly==6.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyodbc==5.2.0
pyparsing==3.2.3
pystray==0.19.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.43
streamlit==1.49.1
streamlit-cookies-manager==0.2.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
